requirements.txt
requests==2.31.0
lxml==4.9.3
fake-useragent==1.2.1
dnspython==2.4.2